restructure tests a bit and implement geometric functions
